BaseGenerator did not cleaned kept data correctly.
BaseGenerator adds the hability to backup some data on which nodes
and edges are created. However when creating a second graph with the
same generator, this data must be cleaned up. This has been added
in a default "end()" method. Each generator has been modified to
call this super.end() method.